+++ dev/examples/gesturesActive.tsx
@@ -1,6 +1,5 @@
 import * as React from "react"
 import { motion } from "@framer"
-import { useAnimation } from "../../src"
 
 const style = {
     width: 100,
@@ -9,18 +8,10 @@ const style = {
 }
 
 export const App = () => {
-    const animation = useAnimation()
     return (
         <motion.div
-            animate={animation}
             whileHover={{ scale: 1.5 }}
             whileTap={{ scale: 0.5, backgroundColor: "rgba(0, 255, 0, .5)" }}
-            onHoverStart={() => {
-                animation.start({ background: "#00F" })
-            }}
-            onHoverEnd={() => {
-                animation.start({ background: "#F00" })
-            }}
             style={style}
         />
     )