+++ src/behaviours/index.ts
@@ -1,1 +1,0 @@
-export { useDraggable, DraggableProps, DragHandlers } from "./use-draggable"