+++ src/gestures/__tests__/tap.test.tsx
@@ -4,7 +4,7 @@ import { render } from "@testing-library/react"
 import { fireEvent } from "@testing-library/dom"
 import { motionValue } from "../../value"
 import { mouseEnter, mouseLeave } from "../../../jest.setup"
-import { drag, MockDrag } from "../../behaviours/__tests__/index.test"
+import { drag, MockDrag } from "../../behaviours/__tests__/utils"
 import sync from "framesync"
 
 function mockWhenFirstArgumentIs(
@@ -91,9 +91,7 @@ describe("tap", () => {
         expect(mockedRemoveMouseUpListener).toHaveBeenCalledTimes(0)
         unmount()
         expect(mockedAddMouseUpListener).toHaveBeenCalledTimes(1)
-        // When unmounting, both use-pan-gesture and use-tap-gesture remove their mouseup listener
-        // That's why this is called two times instead of one
-        expect(mockedRemoveMouseUpListener).toHaveBeenCalledTimes(2)
+        expect(mockedRemoveMouseUpListener).toHaveBeenCalledTimes(1)
 
         expect(tap).toBeCalledTimes(0)
     })