+++ src/gestures/use-tap-gesture.ts
@@ -1,17 +1,12 @@
-import { RefObject, useEffect, useMemo, useRef } from "react"
-import {
-    usePointerEvents,
-    useConditionalPointerEvents,
-    EventInfo,
-    Point,
-    EventHandler,
-} from "../events"
+import { RefObject, useRef } from "react"
+import { EventInfo, Point, EventHandler } from "../events/types"
 import { TargetAndTransition } from "../types"
 import { isNodeOrChild } from "./utils/is-node-or-child"
 import { getGesturePriority } from "./utils/gesture-priority"
-import { ControlsProp } from "./types"
-import { safeWindow } from "../events/utils/window"
+import { ControlsProp, RemoveEvent } from "./types"
 import { getGlobalLock } from "../behaviours/utils/lock"
+import { addPointerEvent, usePointerEvent } from "../events/use-pointer-event"
+import { useUnmountEffect } from "../utils/use-unmount-effect"
 
 const tapGesturePriority = getGesturePriority("whileTap")
 
@@ -184,151 +179,93 @@ export interface TapHandlers {
 }
 
 /**
- * Explanation:
- *
- * Currently there's a problem where props can change while a gesture is active. In
- * this hook, we have event handlers that are created in `useMemo`, which is sync.
- * The event handlers themselves are activated in `useEvent` with `useEffect`, which
- * is async.
- *
- * Most event handlers are being set like this: onTap={() => {}}
- * This creates a new function every render, thus breaking `useMemo`. When any props
- * are set during a gesture, this creates a new event. We then have a `useEffect` in this
- * hook that, when props change, unsets the previous event handlers. I suspect that at
- * some point, the pointer up event is fired in between the previous one
- * being deactivate and the new one being activated.
- *
- * I attempted to rewrite the hook without `useMemo` at all but encountered the same issue.
- * At some point the `pointerUp` event is being unset and not reapplied until after it fires.
- *
- * By setting props to a ref we maintain longer gesture sessions and don't encounter this funny
- * business, but I'd rather a solution that required less code.
- *
- * @param props
- */
-const usePropsRef = <T>(props: T) => {
-    const propsRef = useRef(props)
-
-    for (const key in props) {
-        propsRef.current[key] = props[key]
-    }
-
-    return propsRef.current
-}
-
-/**
  * @param handlers -
  * @internal
  */
 export function useTapGesture(
-    handlers: TapHandlers & ControlsProp
-): { onPointerDown: EventHandler }
-export function useTapGesture(
-    handlers: TapHandlers & ControlsProp,
+    {
+        onTap,
+        onTapStart,
+        onTapCancel,
+        whileTap,
+        controls,
+    }: TapHandlers & ControlsProp,
     ref: RefObject<Element>
-): undefined
-export function useTapGesture(
-    props: TapHandlers & ControlsProp,
-    ref?: RefObject<Element>
-): undefined | { onPointerDown: EventHandler } {
+) {
+    const hasTapListeners = onTap || onTapStart || onTapCancel || whileTap
     const isTapping = useRef(false)
-    const { onTap, onTapStart, onTapCancel, whileTap, controls } = props
-    const propsRef = usePropsRef(props)
+
+    const cancelPointerEventListener = useRef<RemoveEvent | undefined | null>(
+        null
+    )
+
+    function removePointerUp() {
+        cancelPointerEventListener.current &&
+            cancelPointerEventListener.current()
+        cancelPointerEventListener.current = null
+    }
 
     if (whileTap && controls) {
         controls.setOverride(whileTap, tapGesturePriority)
     }
 
-    const handlers = useMemo(
-        () => {
-            if (!onTap && !onTapStart && !onTapCancel && !whileTap) {
-                return {
-                    onPointerUp: () => {},
-                    onPointerDown: () => {},
-                }
-            }
+    // We load this event handler into a ref so we can later refer to
+    // onPointerUp.current which will always have reference to the latest props
+    const onPointerUp = useRef<EventHandler | null>(null)
+    onPointerUp.current = (event, info) => {
+        const element = ref.current
 
-            const onPointerUp = (
-                event: MouseEvent | TouchEvent | PointerEvent,
-                { point }: EventInfo
-            ) => {
-                stopPointerUp()
-                if (!isTapping.current) {
-                    return
-                }
+        removePointerUp()
+        if (!isTapping.current || !element) return
 
-                isTapping.current = false
+        isTapping.current = false
 
-                if (controls && propsRef.whileTap) {
-                    controls.clearOverride(tapGesturePriority)
-                }
+        if (controls && whileTap) {
+            controls.clearOverride(tapGesturePriority)
+        }
 
-                // Check the gesture lock - if we get it, it means no drag gesture is active
-                // and we can safely fire the tap gesture.
-                const openGestureLock = getGlobalLock(true)
+        // Check the gesture lock - if we get it, it means no drag gesture is active
+        // and we can safely fire the tap gesture.
+        const openGestureLock = getGlobalLock(true)
+        if (!openGestureLock) return
+        openGestureLock()
 
-                if (!openGestureLock) {
-                    return
-                }
-                openGestureLock()
+        if (!isNodeOrChild(element, event.target as Element)) {
+            onTapCancel && onTapCancel(event, info)
+        } else {
+            onTap && onTap(event, info)
+        }
+    }
 
-                if (
-                    !ref ||
-                    !ref.current ||
-                    !isNodeOrChild(ref.current, event.target as Element)
-                ) {
-                    if (propsRef.onTapCancel) {
-                        propsRef.onTapCancel(event, { point })
-                    }
-                    return
-                }
+    function onPointerDown(
+        event: MouseEvent | TouchEvent | PointerEvent,
+        info: EventInfo
+    ) {
+        removePointerUp()
 
-                if (propsRef.onTap) {
-                    propsRef.onTap(event, { point })
-                }
-            }
+        cancelPointerEventListener.current = addPointerEvent(
+            window,
+            "pointerup",
+            (event, info) => onPointerUp.current!(event, info)
+        )
 
-            const onPointerDown = (
-                event: MouseEvent | TouchEvent | PointerEvent,
-                { point }: EventInfo
-            ) => {
-                startPointerUp()
-                if (!ref || !ref.current) return
-                isTapping.current = true
+        const element = ref.current
+        if (!element || isTapping.current) return
 
-                if (propsRef.onTapStart) {
-                    propsRef.onTapStart(event, { point })
-                }
+        isTapping.current = true
 
-                if (controls && propsRef.whileTap) {
-                    controls.startOverride(tapGesturePriority)
-                }
-            }
+        onTapStart && onTapStart(event, info)
 
-            return { onPointerUp, onPointerDown }
-        },
-        [
-            onTap === undefined,
-            onTapStart === undefined,
-            onTapCancel === undefined,
-        ]
+        if (controls && whileTap) {
+            controls.startOverride(tapGesturePriority)
+        }
+    }
+
+    usePointerEvent(
+        ref,
+        "pointerdown",
+        hasTapListeners ? onPointerDown : undefined
     )
 
-    const [startPointerUp, stopPointerUp] = usePointerEvents(
-        { onPointerUp: handlers.onPointerUp },
-        safeWindow
-    )
-    useEffect(
-        () => {
-            return () => {
-                stopPointerUp()
-            }
-        },
-        [handlers.onPointerUp]
-    )
-
-    return useConditionalPointerEvents(
-        { onPointerDown: handlers.onPointerDown },
-        ref
-    )
+    useUnmountEffect(removePointerUp)
 }