+++ src/index.ts
@@ -46,7 +46,7 @@ export {
     VariantLabels,
 } from "./motion/types"
 export { GestureHandlers } from "./gestures"
-export { DraggableProps, DragHandlers } from "./behaviours"
+export { DraggableProps, DragHandlers } from "./behaviours/types"
 export {
     Orchestration,
     Tween,
@@ -68,8 +68,7 @@ export {
     Variant,
     Variants,
 } from "./types"
-export { EventInfo } from "./events"
-export { safeWindow } from "./events/utils/window"
+export { EventInfo } from "./events/types"
 export { AnimatePresenceProps } from "./components/AnimatePresence/types"
 export { AnimatePresence } from "./components/AnimatePresence"
 export { UnstableSyncLayout } from "./components/SyncLayout"